requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.7
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.6.0
distlib==0.3.0
Django==3.1.1
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-taggit==1.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
idna==2.10
install==1.3.4
oauthlib==3.1.0
Pillow==8.3.1
pipenv==2018.11.26
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-framework-generic-relations==2.0.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.2
urllib3==1.25.10
virtualenv==20.0.21
virtualenv-clone==0.5.4
